ACC3 deployment on 21Sep

Related work items: #9699, #16671, #16679, #17141, #17326, #17330, #17843, #18626, #18838

+++ dotnet/corecomponents/net.atos.daf.ct2.reports/entity/AlertThresholdDetails.cs
@@ -8,7 +8,7 @@ namespace net.atos.daf.ct2.reports.entity
     {
         public int AlertId { get; set; }
         public string AlertLevel { get; set; }
-        public int ThresholdValue { get; set; }
+        public double ThresholdValue { get; set; }
         public string ThresholdUnit { get; set; }
 
     }
